+++ src/models/IGetTransactionsToApproveRequest.ts
@@ -7,4 +7,14 @@ export interface IGetTransactionsToApproveRequest {
      * Number of bundles to go back to determine the transactions for approval.
      */
     depth: number;
+
+    /**
+     * The reference to include in the lookup.
+     */
+    reference?: string;
+
+    /**
+     * The number of walks to perform.
+     */
+    numWalks?: number;
 }